Adding 21 game and 21 game with menu
theGame21.py is the finished standalone game of 21. 21menuExample.py is an example of a menu I found online that we could implement (The python package is PyQt5). Also the zip file of the cards are attached (in the crd hash map in the code, make sure to change the file path names to where you place the folder)