User Interface Example
This is an example of a user interface menu that does not have the message scrolling in the terminal saying "The event loop is already running". NOTE: comment out the line main() at the very last line of the file theGame21.py before running this menu file, so the program does not run before you click on the button.