Adding files for 21 game: keyboard/voice
Fixed a game rule in bareBones21.py where the player should win right away if they are dealt 21 in the beginning of the game (should not be asked to hit or stand). bareBones21Audio.py is the 21 game with solely audio. I also fixed the game logic in hand_score where the ace hand score changes (1 or 11) per game rules.